+++ js_modules/dagit/src/execution/types/PipelineRunFragment.ts
@@ -65,7 +65,7 @@ export interface PipelineRunFragment_executionPlan_steps_solid {
 export interface PipelineRunFragment_executionPlan_steps {
   name: string;
   solid: PipelineRunFragment_executionPlan_steps_solid;
-  tag: StepTag;
+  kind: StepKind;
 }
 
 export interface PipelineRunFragment_executionPlan {
@@ -109,7 +109,7 @@ export enum PipelineRunStatus {
   SUCCESS = "SUCCESS",
 }
 
-export enum StepTag {
+export enum StepKind {
   INPUT_EXPECTATION = "INPUT_EXPECTATION",
   INPUT_THUNK = "INPUT_THUNK",
   JOIN = "JOIN",