+++ python_modules/dagit/dagit/schema/execution.py
@@ -62,9 +62,9 @@ class DauphinExecutionStepInput(dauphin.ObjectType):
         return info.schema.type_named('ExecutionStep')(self._step_input.prev_output_handle.step)
 
 
-class DauphinStepTag(dauphin.Enum):
+class DauphinStepKind(dauphin.Enum):
     class Meta:
-        name = 'StepTag'
+        name = 'StepKind'
 
     TRANSFORM = 'TRANSFORM'
     INPUT_EXPECTATION = 'INPUT_EXPECTATION'
@@ -78,27 +78,27 @@ class DauphinStepTag(dauphin.Enum):
     def description(self):
         # self ends up being the internal class "EnumMeta" in dauphin
         # so we can't do a dictionary lookup which is awesome
-        if self == DauphinStepTag.TRANSFORM:
+        if self == DauphinStepKind.TRANSFORM:
             return 'This is the user-defined transform step'
-        elif self == DauphinStepTag.INPUT_EXPECTATION:
+        elif self == DauphinStepKind.INPUT_EXPECTATION:
             return 'Expectation defined on an input'
-        elif self == DauphinStepTag.OUTPUT_EXPECTATION:
+        elif self == DauphinStepKind.OUTPUT_EXPECTATION:
             return 'Expectation defined on an output'
-        elif self == DauphinStepTag.JOIN:
+        elif self == DauphinStepKind.JOIN:
             return (
                 'Sometimes we fan out compute on identical values (e.g. multiple expectations '
                 'in parallel). We synthesize these in a join step to consolidate to a single '
                 'output that the next computation can depend on.'
             )
-        elif self == DauphinStepTag.SERIALIZE:
+        elif self == DauphinStepKind.SERIALIZE:
             return (
                 'This is a special system-defined step to serialize an intermediate value if '
                 'the pipeline is configured to do that.'
             )
 
-        elif self == DauphinStepTag.INPUT_THUNK:
+        elif self == DauphinStepKind.INPUT_THUNK:
             return 'Special system-defined step to represent an input specified in the environment'
-        elif self == DauphinStepTag.MATERIALIZATION_THUNK:
+        elif self == DauphinStepKind.MATERIALIZATION_THUNK:
             return (
                 'Special system-defined step to represent an output materialization specified in '
                 'the environment'
@@ -115,7 +115,7 @@ class DauphinExecutionStep(dauphin.ObjectType):
     inputs = dauphin.non_null_list('ExecutionStepInput')
     outputs = dauphin.non_null_list('ExecutionStepOutput')
     solid = dauphin.NonNull('Solid')
-    tag = dauphin.NonNull('StepTag')
+    kind = dauphin.NonNull('StepKind')
 
     def __init__(self, execution_step):
         super(DauphinExecutionStep, self).__init__()
@@ -139,5 +139,5 @@ class DauphinExecutionStep(dauphin.ObjectType):
     def resolve_solid(self, info):
         return info.schema.type_named('Solid')(self.execution_step.solid)
 
-    def resolve_tag(self, _info):
-        return self.execution_step.tag
+    def resolve_kind(self, _info):
+        return self.execution_step.kind