+++ python_modules/dagster/dagster/core/execution_plan/objects.py
@@ -1,10 +1,13 @@
 from collections import namedtuple
+from contextlib import contextmanager
+import copy
 from enum import Enum
 import toposort
 
 from dagster import check
+from dagster.utils import merge_dicts
 from dagster.core.system_config.objects import EnvironmentConfig
-from dagster.core.definitions import PipelineDefinition, Solid
+from dagster.core.definitions import PipelineDefinition, Solid, SolidOutputHandle
 from dagster.core.errors import DagsterError
 from dagster.core.execution_context import RuntimeExecutionContext
 from dagster.core.types.runtime import RuntimeType
@@ -51,29 +54,29 @@ class StepFailureData(namedtuple('_StepFailureData', 'dagster_error')):
         )
 
 
-class StepResult(namedtuple('_StepResult', 'success step tag success_data failure_data')):
+class StepResult(namedtuple('_StepResult', 'success step kind success_data failure_data')):
     @staticmethod
-    def success_result(step, tag, success_data):
+    def success_result(step, kind, success_data):
         return StepResult(
             success=True,
             step=check.inst_param(step, 'step', ExecutionStep),
-            tag=check.inst_param(tag, 'tag', StepTag),
+            kind=check.inst_param(kind, 'kind', StepKind),
             success_data=check.inst_param(success_data, 'success_data', StepSuccessData),
             failure_data=None,
         )
 
     @staticmethod
-    def failure_result(step, tag, failure_data):
+    def failure_result(step, kind, failure_data):
         return StepResult(
             success=False,
             step=check.inst_param(step, 'step', ExecutionStep),
-            tag=check.inst_param(tag, 'tag', StepTag),
+            kind=check.inst_param(kind, 'kind', StepKind),
             success_data=None,
             failure_data=check.inst_param(failure_data, 'failure_data', StepFailureData),
         )
 
 
-class StepTag(Enum):
+class StepKind(Enum):
     TRANSFORM = 'TRANSFORM'
     INPUT_EXPECTATION = 'INPUT_EXPECTATION'
     OUTPUT_EXPECTATION = 'OUTPUT_EXPECTATION'
@@ -108,10 +111,15 @@ class StepOutput(namedtuple('_StepOutput', 'name runtime_type')):
 class ExecutionStep(
     namedtuple(
         '_ExecutionStep',
-        'key step_inputs step_input_dict step_outputs step_output_dict compute_fn tag solid',
+        'key step_inputs step_input_dict step_outputs step_output_dict compute_fn kind solid tags',
     )
 ):
-    def __new__(cls, key, step_inputs, step_outputs, compute_fn, tag, solid):
+    def __new__(cls, key, step_inputs, step_outputs, compute_fn, kind, solid, tags):
+        check.param_invariant('pipeline' in tags, 'tags', 'Step must have pipeline tag')
+        check.param_invariant('solid' in tags, 'tags', 'Step must have solid tag')
+        check.param_invariant('solid_definition' in tags, 'tags', 'Step must have solid tag')
+        check.dict_param(tags, 'tags', key_type=str, value_type=str)
+
         return super(ExecutionStep, cls).__new__(
             cls,
             key=check.str_param(key, 'key'),
@@ -120,9 +128,22 @@ class ExecutionStep(
             step_outputs=check.list_param(step_outputs, 'step_outputs', of_type=StepOutput),
             step_output_dict={so.name: so for so in step_outputs},
             compute_fn=check.callable_param(compute_fn, 'compute_fn'),
-            tag=check.inst_param(tag, 'tag', StepTag),
+            kind=check.inst_param(kind, 'kind', StepKind),
             solid=check.inst_param(solid, 'solid', Solid),
+            tags=merge_dicts({'step_key': key}, tags),
         )
+
+    @property
+    def pipeline_name(self):
+        return self.tags['pipeline']
+
+    @property
+    def solid_name(self):
+        return self.tags['solid']
+
+    @property
+    def solid_definition_name(self):
+        return self.tags['solid_definition']
 
     def __getnewargs__(self):
         # print('getnewargs was called')
@@ -131,7 +152,7 @@ class ExecutionStep(
             self.step_inputs,
             self.step_outputs,
             self.compute_fn,
-            self.tag,
+            self.kind,
             self.solid,
         )
 
@@ -141,8 +162,9 @@ class ExecutionStep(
             step_inputs=step_inputs,
             step_outputs=self.step_outputs,
             compute_fn=self.compute_fn,
-            tag=self.tag,
+            kind=self.kind,
             solid=self.solid,
+            tags=self.tags,
         )
 
     def has_step_output(self, name):
@@ -230,3 +252,42 @@ class ExecutionPlanSubsetInfo(namedtuple('_ExecutionPlanSubsetInfo', 'subset inp
             set(check.list_param(included_steps, 'included_steps', of_type=str)),
             check.opt_dict_param(inputs, 'inputs'),
         )
+
+
+class StepOutputMap(dict):
+    def __getitem__(self, key):
+        check.inst_param(key, 'key', SolidOutputHandle)
+        return dict.__getitem__(self, key)
+
+    def __setitem__(self, key, val):
+        check.inst_param(key, 'key', SolidOutputHandle)
+        check.inst_param(val, 'val', StepOutputHandle)
+        return dict.__setitem__(self, key, val)
+
+
+# This is the state that is built up during the execution plan build process.
+# steps is just a list of the steps that have been created
+# step_output_map maps logical solid outputs (solid_name, output_name) to particular
+# step outputs. This covers the case where a solid maps to multiple steps
+# and one wants to be able to attach to the logical output of a solid during execution
+class StepBuilderState:
+    def __init__(self, pipeline_name, initial_tags=None):
+        self.steps = []
+        self.step_output_map = StepOutputMap()
+        self._tags = check.opt_dict_param(
+            initial_tags, 'initial_tags', key_type=str, value_type=str
+        )
+        self._tags['pipeline'] = pipeline_name
+
+    def get_tags(self, **additional_tags):
+        additional_tags.update(self._tags)
+        return additional_tags
+
+    @contextmanager
+    def push_tags(self, **kwargs):
+        old_tags = copy.copy(self._tags)
+        self._tags.update(kwargs)
+        try:
+            yield
+        finally:
+            self._tags = old_tags