+++ python_modules/dagster/dagster/utils/test.py
@@ -19,9 +19,9 @@ from dagster.core.execution_context import RuntimeExecutionContext
 from dagster.core.utility_solids import define_stub_solid
 
 
-def create_test_runtime_execution_context(loggers=None, resources=None):
+def create_test_runtime_execution_context(loggers=None, resources=None, tags=None):
     run_id = str(uuid.uuid4())
-    return RuntimeExecutionContext(run_id, loggers, resources)
+    return RuntimeExecutionContext(run_id=run_id, loggers=loggers, resources=resources, tags=tags)
 
 
 def _unlink_swallow_errors(path):