+++ CHANGES.rst
@@ -11,16 +11,18 @@
       - #670: Internal system error "dagster.check.CheckError: Invariant failed. Description: Should not be in context" raised when user throwing error during transform. Now the appropriate user error should be raised.
       - #672: Dagit sometimes hangs (TypeError: unsupported operand type(s) for -: 'float' and 'NoneType' in console log)
 
-   - API Additions
+   - API Additions and Changes
       - New decorated-based @resource API as a more concise alternative to ResourceDefinition
       - Dagster config type system now supports enum types. (dagster.Enum and dagster.EnumType) 
       - New top level properties `resources` and `log` on info.
+      - The context stack in RuntimeExecutionContext is no longer modify-able by the user during a transform. It has been renamed to 'tags'.
+      - ReentrantInfo has been renamed to ExecutionMetadata
 
-   - Notes
+   - GraphQL Schema Changes
       - GraphQL queries and mutations taking a pipeline name now take both a pipeline name and an optional
         solid subset and have slightly improved call signatures.
       - The config and runtime type system split is now reflected in the GraphQL frontend. This was the infrastructure
         piece that allowed the fix to #598. runtimeTypeOrError, configTypeOrError are now top level fields, and there
         are configTypes and runtimeTypes fields on Pipeline. Top-level field type and types property on Pipeline has
         been eliminated.
-
+      - StepTag has been renamed to StepKind